+++ flow/flow.go
@@ -25,8 +25,9 @@ import (
 	"net/http"
 
 	"github.com/Sirupsen/logrus"
-	"github.com/northwesternmutual/kanali/controller"
 	"github.com/northwesternmutual/kanali/metrics"
+	"github.com/northwesternmutual/kanali/spec"
+	"github.com/northwesternmutual/kanali/tracer"
 	"github.com/opentracing/opentracing-go"
 )
 
@@ -41,12 +42,12 @@ func (f *Flow) Add(steps ...Step) {
 }
 
 // Play executes all step in a flow in the order they were added.
-func (f *Flow) Play(ctx context.Context, metrics *metrics.Metrics, ctlr *controller.Controller, w http.ResponseWriter, r *http.Request, resp *http.Response, trace opentracing.Span) error {
+func (f *Flow) Play(ctx context.Context, proxy *spec.APIProxy, metrics *metrics.Metrics, w http.ResponseWriter, r *http.Request, resp *http.Response, trace opentracing.Span) error {
 	logrus.Debugf("flow with %d step about to play", len(*f))
 	for _, step := range *f {
 		logrus.Debugf("playing step %s", step.GetName())
-		if err := step.Do(ctx, metrics, ctlr, w, r, resp, trace); err != nil {
-			trace.SetTag("error", true)
+		if err := step.Do(ctx, proxy, metrics, w, r, resp, trace); err != nil {
+			trace.SetTag(tracer.Error, true)
 			trace.LogKV(
 				"event", "error",
 				"error.message", err.Error(),
@@ -55,5 +56,4 @@ func (f *Flow) Play(ctx context.Context, metrics *metrics.Metrics, ctlr *control
 		}
 	}
 	return nil
-
 }