+++ handlers/handler.go
@@ -30,20 +30,18 @@ import (
 	"time"
 
 	"github.com/Sirupsen/logrus"
-	"github.com/northwesternmutual/kanali/config"
-	"github.com/northwesternmutual/kanali/controller"
 	"github.com/northwesternmutual/kanali/metrics"
 	"github.com/northwesternmutual/kanali/monitor"
+	"github.com/northwesternmutual/kanali/spec"
+	"github.com/northwesternmutual/kanali/tracer"
 	"github.com/northwesternmutual/kanali/utils"
 	"github.com/opentracing/opentracing-go"
-	"github.com/spf13/viper"
 )
 
 // Handler is used to provide additional parameters to an HTTP handler
 type Handler struct {
-	*controller.Controller
 	*monitor.InfluxController
-	H func(ctx context.Context, m *metrics.Metrics, c *controller.Controller, w http.ResponseWriter, r *http.Request, trace opentracing.Span) error
+	H func(ctx context.Context, proxy *spec.APIProxy, m *metrics.Metrics, w http.ResponseWriter, r *http.Request, trace opentracing.Span) error
 }
 
 func (h Handler) serveHTTP(w http.ResponseWriter, r *http.Request) {
@@ -69,98 +67,78 @@ func (h Handler) serveHTTP(w http.ResponseWriter, r *http.Request) {
 		}()
 	}()
 
-	// start a global trace
 	sp := opentracing.StartSpan(fmt.Sprintf("%s %s",
 		r.Method,
 		r.URL.EscapedPath(),
 	))
-
-	closer, str, err := utils.DupReaderAndString(r.Body)
-	if err != nil {
-		logrus.Errorf("error copying request body: %s", err.Error())
-	}
-
-	// copy new reader into freshly drained reader
-	r.Body = closer
-
-	sp.SetTag("http.request_body", str)
-	sp.SetTag("http.url", r.URL.EscapedPath())
-	sp.SetTag("http.method", r.Method)
-
-	jsonHeaders, err := json.Marshal(utils.FlattenHTTPHeaders(utils.OmitHeaderValues(r.Header, viper.GetString(config.FlagProxyHeaderMaskValue.GetLong()), viper.GetStringSlice(config.FlagProxyMaskHeaderKeys.GetLong())...)))
-	if err != nil {
-		logrus.Warnf("could not marsah request headers into JSON - tracing data maybe not be as expected")
-	} else {
-		sp.SetTag("http.headers", string(jsonHeaders))
-	}
-
 	defer sp.Finish()
 
-	err = h.H(context.Background(), m, h.Controller, w, r, sp)
+	tracer.HydrateSpanFromRequest(r, sp)
 
-	// handle request errors
-	if err != nil {
+	err := h.H(context.Background(), &spec.APIProxy{}, m, w, r, sp)
+	if err == nil {
+		return
+	}
 
-		// all errors will need the application/json Content-Type header
-		w.Header().Set("Content-Type", "application/json")
+	// all errors will need the application/json Content-Type header
+	w.Header().Set("Content-Type", "application/json")
 
-		// we'll have multiple types off errors
-		switch e := err.(type) {
-		case utils.Error:
+	// we'll have multiple types off errors
+	switch e := err.(type) {
+	case utils.Error:
 
-			sp.SetTag("http.status_code", e.Status())
+		sp.SetTag(tracer.HTTPResponseStatusCode, e.Status())
 
-			// log error
-			logrus.WithFields(logrus.Fields{
-				"method": r.Method,
-				"uri":    r.URL.EscapedPath(),
-			}).Error(e.Error())
+		// log error
+		logrus.WithFields(logrus.Fields{
+			"method": r.Method,
+			"uri":    r.URL.EscapedPath(),
+		}).Error(e.Error())
 
-			m.Add(metrics.Metric{Name: "http_response_code", Value: strconv.Itoa(e.Status()), Index: true})
+		m.Add(metrics.Metric{Name: "http_response_code", Value: strconv.Itoa(e.Status()), Index: true})
 
-			errStatus, err := json.Marshal(utils.JSONErr{Code: e.Status(), Msg: e.Error()})
-			if err != nil {
-				logrus.Warnf("could not marsah request headers into JSON - tracing data maybe not be as expected")
-			} else {
-				sp.SetTag("http.response_body", string(errStatus))
-			}
-
-			// write error code to response
-			w.WriteHeader(e.Status())
-
-			// write error message to response
-			if err := json.NewEncoder(w).Encode(utils.JSONErr{Code: e.Status(), Msg: e.Error()}); err != nil {
-				logrus.Fatal(err.Error())
-			}
-
-		default:
-
-			sp.SetTag("http.status_code", http.StatusInternalServerError)
-
-			// log error
-			logrus.WithFields(logrus.Fields{
-				"method": r.Method,
-				"uri":    r.URL.EscapedPath(),
-			}).Error("unknown error")
-
-			m.Add(metrics.Metric{Name: "http_response_code", Value: strconv.Itoa(http.StatusInternalServerError), Index: true})
-
-			errStatus, err := json.Marshal(utils.JSONErr{Code: http.StatusInternalServerError, Msg: "unknown error"})
-			if err != nil {
-				logrus.Warnf("could not marsah request headers into JSON - tracing data maybe not be as expected")
-			} else {
-				sp.SetTag("http.response_body", string(errStatus))
-			}
-
-			// write error code to response
-			w.WriteHeader(http.StatusInternalServerError)
-
-			// write error message to response
-			if err := json.NewEncoder(w).Encode(utils.JSONErr{Code: http.StatusInternalServerError, Msg: "unknown error"}); err != nil {
-				logrus.Fatal(err.Error())
-			}
-
+		errStatus, err := json.Marshal(utils.JSONErr{Code: e.Status(), Msg: e.Error()})
+		if err != nil {
+			logrus.Warnf("could not marsah request headers into JSON - tracing data maybe not be as expected")
+		} else {
+			sp.SetTag(tracer.HTTPResponseBody, string(errStatus))
 		}
+
+		// write error code to response
+		w.WriteHeader(e.Status())
+
+		// write error message to response
+		if err := json.NewEncoder(w).Encode(utils.JSONErr{Code: e.Status(), Msg: e.Error()}); err != nil {
+			logrus.Fatal(err.Error())
+		}
+
+	default:
+
+		sp.SetTag(tracer.HTTPResponseStatusCode, http.StatusInternalServerError)
+
+		// log error
+		logrus.WithFields(logrus.Fields{
+			"method": r.Method,
+			"uri":    r.URL.EscapedPath(),
+		}).Error("unknown error")
+
+		m.Add(metrics.Metric{Name: "http_response_code", Value: strconv.Itoa(http.StatusInternalServerError), Index: true})
+
+		errStatus, err := json.Marshal(utils.JSONErr{Code: http.StatusInternalServerError, Msg: "unknown error"})
+		if err != nil {
+			logrus.Warnf("could not marsah request headers into JSON - tracing data maybe not be as expected")
+		} else {
+			sp.SetTag(tracer.HTTPResponseBody, string(errStatus))
+		}
+
+		// write error code to response
+		w.WriteHeader(http.StatusInternalServerError)
+
+		// write error message to response
+		if err := json.NewEncoder(w).Encode(utils.JSONErr{Code: http.StatusInternalServerError, Msg: "unknown error"}); err != nil {
+			logrus.Fatal(err.Error())
+		}
+
 	}
 }