+++ steps/validateproxy.go
@@ -26,9 +26,9 @@ import (
 	"net/http"
 
 	"github.com/Sirupsen/logrus"
-	"github.com/northwesternmutual/kanali/controller"
 	"github.com/northwesternmutual/kanali/metrics"
 	"github.com/northwesternmutual/kanali/spec"
+	"github.com/northwesternmutual/kanali/tracer"
 	"github.com/northwesternmutual/kanali/utils"
 	"github.com/opentracing/opentracing-go"
 )
@@ -44,7 +44,7 @@ func (step ValidateProxyStep) GetName() string {
 }
 
 // Do executes the logic of the ValidateProxyStep step
-func (step ValidateProxyStep) Do(ctx context.Context, m *metrics.Metrics, c *controller.Controller, w http.ResponseWriter, r *http.Request, resp *http.Response, trace opentracing.Span) error {
+func (step ValidateProxyStep) Do(ctx context.Context, proxy *spec.APIProxy, m *metrics.Metrics, w http.ResponseWriter, r *http.Request, resp *http.Response, trace opentracing.Span) error {
 
 	untypedProxy, err := spec.ProxyStore.Get(r.URL.EscapedPath())
 	if err != nil || untypedProxy == nil {
@@ -52,8 +52,8 @@ func (step ValidateProxyStep) Do(ctx context.Context, m *metrics.Metrics, c *con
 			logrus.Error(err.Error())
 		}
 
-		trace.SetTag("kanali.proxy_name", "unknown")
-		trace.SetTag("kanali.proxy_namespace", "unknown")
+		trace.SetTag(tracer.KanaliProxyName, "unknown")
+		trace.SetTag(tracer.KanaliProxyNamespace, "unknown")
 
 		m.Add(
 			metrics.Metric{Name: "proxy_name", Value: "unknown", Index: true},
@@ -63,22 +63,12 @@ func (step ValidateProxyStep) Do(ctx context.Context, m *metrics.Metrics, c *con
 		return utils.StatusError{Code: http.StatusNotFound, Err: errors.New("proxy not found")}
 	}
 
-	proxy, ok := untypedProxy.(spec.APIProxy)
-	if !ok {
+	typedProxy, _ := untypedProxy.(spec.APIProxy)
 
-		trace.SetTag("kanali.proxy_name", "unknown")
-		trace.SetTag("kanali.proxy_namespace", "unknown")
+	*proxy = typedProxy
 
-		m.Add(
-			metrics.Metric{Name: "proxy_name", Value: "unknown", Index: true},
-			metrics.Metric{Name: "proxy_namespace", Value: "unknown", Index: true},
-		)
-
-		return utils.StatusError{Code: http.StatusNotFound, Err: errors.New("proxy not found")}
-	}
-
-	trace.SetTag("kanali.proxy_name", proxy.ObjectMeta.Name)
-	trace.SetTag("kanali.proxy_namespace", proxy.ObjectMeta.Namespace)
+	trace.SetTag(tracer.KanaliProxyName, proxy.ObjectMeta.Name)
+	trace.SetTag(tracer.KanaliProxyNamespace, proxy.ObjectMeta.Namespace)
 
 	m.Add(
 		metrics.Metric{Name: "proxy_name", Value: proxy.ObjectMeta.Name, Index: true},