+++ tracer/jaeger_test.go
@@ -18,18 +18,32 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package utils
+package tracer
 
 import (
-	"encoding/base64"
+	"bytes"
 	"testing"
 
+	"github.com/Sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 )
 
-func TestDecodeBase64ByteArray(t *testing.T) {
-	arr1, _ := DecodeBase64ByteArray([]byte(base64.StdEncoding.EncodeToString([]byte("abc123"))))
-	assert.Equal(t, []byte("abc123"), arr1)
-	_, err := DecodeBase64ByteArray([]byte("notbase64"))
-	assert.NotNil(t, err)
+func TestCustomLoggerError(t *testing.T) {
+	logger := customLogger{}
+
+	writerOne := new(bytes.Buffer)
+	writerTwo := new(bytes.Buffer)
+	logrus.SetOutput(writerOne)
+	logger.Error("custom error message")
+	logrus.SetOutput(writerTwo)
+	logrus.Error("custom error message")
+	assert.Equal(t, writerOne.String(), writerTwo.String())
+
+	writerOne = new(bytes.Buffer)
+	writerTwo = new(bytes.Buffer)
+	logrus.SetOutput(writerOne)
+	logger.Infof("custom %s message", "info")
+	logrus.SetOutput(writerTwo)
+	logrus.Infof("custom %s message", "info")
+	assert.Equal(t, writerOne.String(), writerTwo.String())
 }