+++ utils/utils_test.go
@@ -21,9 +21,6 @@
 package utils
 
 import (
-	"bytes"
-	"io/ioutil"
-	"net/http"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -31,23 +28,19 @@ import (
 )
 
 func TestComputeTargetPath(t *testing.T) {
-
-	assert := assert.New(t)
-
-	assert.Equal("/", ComputeTargetPath("/foo/bar", "", "/foo/bar"), "path not what expected")
-	assert.Equal("/", ComputeTargetPath("/foo/bar", "/", "/foo/bar"), "path not what expected")
-	assert.Equal("/foo", ComputeTargetPath("/foo/bar", "/foo", "/foo/bar"), "path not what expected")
-	assert.Equal("/foo/bar", ComputeTargetPath("/foo/bar", "/foo", "/foo/bar/bar"), "path not what expected")
-	assert.Equal("/bar", ComputeTargetPath("/foo/bar", "", "/foo/bar/bar"), "path not what expected")
-	assert.Equal("/accounts", ComputeTargetPath("/api/v1/example-two", "/", "/api/v1/example-two/accounts"), "path not what expected")
-	assert.Equal("/accounts", ComputeTargetPath("/api/v1/example-two", "/", "/api/v1/example-two/accounts/"), "path not what expected")
-	assert.Equal("/accounts", ComputeTargetPath("/api/v1/example-two", "", "/api/v1/example-two/accounts/"), "path not what expected")
-	assert.Equal("/accounts", ComputeTargetPath("/api/v1/example-two/", "/", "/api/v1/example-two/accounts/"), "path not what expected")
-	assert.Equal("/accounts", ComputeTargetPath("/api/v1/example-two/", "", "/api/v1/example-two/accounts/"), "path not what expected")
-	assert.Equal("/accounts", ComputeTargetPath("/api/v1/example-two/", "", "/api/v1/example-two/accounts"), "path not what expected")
-	assert.Equal("/", ComputeTargetPath("/", "", "/"), "path not what expected")
-	assert.Equal("/", ComputeTargetPath("/", "/", "/"), "path not what expected")
-
+	assert.Equal(t, "/", ComputeTargetPath("/foo/bar", "", "/foo/bar"))
+	assert.Equal(t, "/", ComputeTargetPath("/foo/bar", "/", "/foo/bar"))
+	assert.Equal(t, "/foo", ComputeTargetPath("/foo/bar", "/foo", "/foo/bar"))
+	assert.Equal(t, "/foo/bar", ComputeTargetPath("/foo/bar", "/foo", "/foo/bar/bar"))
+	assert.Equal(t, "/bar", ComputeTargetPath("/foo/bar", "", "/foo/bar/bar"))
+	assert.Equal(t, "/accounts", ComputeTargetPath("/api/v1/example-two", "/", "/api/v1/example-two/accounts"))
+	assert.Equal(t, "/accounts", ComputeTargetPath("/api/v1/example-two", "/", "/api/v1/example-two/accounts/"))
+	assert.Equal(t, "/accounts", ComputeTargetPath("/api/v1/example-two", "", "/api/v1/example-two/accounts/"))
+	assert.Equal(t, "/accounts", ComputeTargetPath("/api/v1/example-two/", "/", "/api/v1/example-two/accounts/"))
+	assert.Equal(t, "/accounts", ComputeTargetPath("/api/v1/example-two/", "", "/api/v1/example-two/accounts/"))
+	assert.Equal(t, "/accounts", ComputeTargetPath("/api/v1/example-two/", "", "/api/v1/example-two/accounts"))
+	assert.Equal(t, "/", ComputeTargetPath("/", "", "/"), "path not what expected")
+	assert.Equal(t, "/", ComputeTargetPath("/", "/", "/"), "path not what expected")
 }
 
 func TestAbsPath(t *testing.T) {
@@ -57,70 +50,6 @@ func TestAbsPath(t *testing.T) {
 	assert.Equal(t, "/foo", p)
 	p, _ = GetAbsPath("//")
 	assert.Equal(t, "", p)
-}
-
-func TestIsValidHTTPMethod(t *testing.T) {
-	assert.False(t, IsValidHTTPMethod("foo"))
-	assert.True(t, IsValidHTTPMethod("GET"))
-	assert.True(t, IsValidHTTPMethod("get"))
-	assert.True(t, IsValidHTTPMethod("POST"))
-	assert.True(t, IsValidHTTPMethod("post"))
-}
-
-func TestFlattenHTTPHeaders(t *testing.T) {
-	h := http.Header{
-		"Foo": []string{"bar"},
-		"Bar": []string{"foo"},
-	}
-	assert.Equal(t, FlattenHTTPHeaders(h), map[string]string{
-		"Foo": "bar",
-		"Bar": "foo",
-	})
-	assert.Nil(t, FlattenHTTPHeaders(nil))
-}
-
-func TestDupReaderAndString(t *testing.T) {
-	closer := ioutil.NopCloser(bytes.NewReader([]byte("test string")))
-	closerTwo, str, _ := DupReaderAndString(closer)
-	assert.Equal(t, str, "test string")
-	data, _ := ioutil.ReadAll(closerTwo)
-	assert.Equal(t, string(data), "test string")
-}
-
-func TestOmitHeaderValues(t *testing.T) {
-
-	assert := assert.New(t)
-
-	h := http.Header{
-		"One":   []string{"two"},
-		"Three": []string{"four"},
-	}
-
-	copy := OmitHeaderValues(h, "omitted", "one")
-	assert.Equal(h, http.Header{
-		"One":   []string{"two"},
-		"Three": []string{"four"},
-	}, "original map should not change")
-	assert.Equal(copy, http.Header{
-		"One":   []string{"omitted"},
-		"Three": []string{"four"},
-	}, "map should be equal")
-
-	copy = OmitHeaderValues(h, "omitted", "one", "foo", "bar")
-	assert.Equal(copy, http.Header{
-		"One":   []string{"omitted"},
-		"Three": []string{"four"},
-	}, "map should be equal")
-
-	copy = OmitHeaderValues(h, "omitted")
-	assert.Equal(copy, http.Header{
-		"One":   []string{"two"},
-		"Three": []string{"four"},
-	}, "original map should not change")
-
-	copy = OmitHeaderValues(nil, "omitted")
-	assert.Nil(copy, "map should be equal")
-
 }
 
 func TestCompareObjectMeta(t *testing.T) {