+++ CHANGELOG.md
@@ -4,18 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.1.5] - 2017-09-14
+## [1.2.0] - 2017-09-24
 ### Added
 - Gzip support.
 - `Update` method to `Store` interface.
+- Additional Opentracing span to capture target request.
+- _Average Target Response Time_ Grafana panel.
 ### Changed
 - Fixed [#61](https://github.com/northwesternmutual/kanali/issues/61)
 - Fixed [#57](https://github.com/northwesternmutual/kanali/issues/57).
-- Using Alpine as base Docker image.
 - Using versioned API key plugin in distributed `Dockefile` and examples.
 - Refactored configuration items.
 - Using `TOML` as configuration format in distribution and examples.
 - Improved test coverage.
+### Removed
+- `controller.Controller` from `Plugin` interface method parameters
 
 ## [1.1.4] - 2017-08-31
 ### Added